+++ GMConverter.UI/Program.cs
@@ -11,10 +11,13 @@ sealed class Program
     [STAThread]
     public static void Main(string[] args)
     {
-        Ab4d.SharpEngine.Licensing.SetLicense(
-            licenseOwner: "David Katz",
-            licenseType: "TrialLicense",
-            license: "7C75-1AE3-E6E9-1175-20B7-4DAE-A684-8E20-55A5-4A97-C0BD-2A71-FDFD-E052-B391-3D14-6DC2");
+        // This free open source license is valid only for the open source project at the following URL:
+        // https://github.com/gmod-workshop/gmconverter
+        // Assembly name: 'GMConverter.UI'
+        // The license is valid for all Ab4d.SharpEngine versions that are published before 2027-05-12.
+        Ab4d.SharpEngine.Licensing.SetLicense(licenseOwner: "David Katz",
+            licenseType: "OpenSourceLicense",
+            license: "A543-105E-0047-F209-CC6E-32CD-D155-4F0F-11AD-2706-9E68-7A5B-4945-D56B-9F96-C0CB-A45F-9339-DB9E-F4CE-C84A-DFAD-82B5-B095-1B");
 
         BuildAvaloniaApp().StartWithClassicDesktopLifetime(args);
     }